refactor: move diagnostics to root module and update changelog
