+++ src/weather_model_graphs/diagnostics.py
